+++ Plugins/Rename/Rename.hpp
@@ -6,6 +6,7 @@
 #include "API/Types.hpp"
 #include "Common.hpp"
 #include "ViewPtr.hpp"
+#include <unordered_set>
 
 using namespace NWNXLib;
 using namespace NWNXLib::API;
@@ -21,9 +22,14 @@ public:
     virtual ~Rename();
 
 private:
+    std::unordered_map<Types::ObjectID, std::unordered_map<Types::ObjectID, std::tuple<CExoString, CExoString, int32_t>>> m_RenamePlayerNames;
+    std::unordered_map<Types::ObjectID, std::tuple<CExoString, CExoString, int32_t>>::iterator m_RenamePlayerNamesIter;
+    std::unordered_map<Types::ObjectID, std::tuple<CExoString, CExoLocString, CExoLocString>> m_RenameOriginalNames;
     int32_t m_RenameOnModuleCharList;
+    std::unordered_set<Types::PlayerID> m_RenameAddedToPlayerList;
     bool m_RenameOnPlayerList;
     bool m_RenameAllowDM;
+    std::string m_RenameAnonymousPlayerName;
 
     static void WriteGameObjUpdate_UpdateObjectHook(
             Services::Hooks::CallType,
@@ -47,6 +53,10 @@ private:
             CNWSMessage*,
             Types::PlayerID,
             CNWSPlayer*);
+    static void SendServerToPlayerDungeonMasterUpdatePartyListHook(
+            NWNXLib::Services::Hooks::CallType,
+            CNWSMessage*,
+            Types::PlayerID);
     static void SendServerToPlayerExamineGui_CreatureDataHook(
             Services::Hooks::CallType,
             CNWSMessage*,
@@ -76,8 +86,11 @@ private:
     CExoLocString ContainString(const std::string& str);
     std::string GenerateRandomPlayerName(size_t length);
 
+    void SendNameUpdate(CNWSCreature *targetCreature, Types::PlayerID observerPlayerId);
+
     ArgumentStack SetPCNameOverride(ArgumentStack&& args);
     ArgumentStack GetPCNameOverride(ArgumentStack&& args);
+    ArgumentStack ClearPCNameOverride(ArgumentStack&& args);
 
     CNWSPlayer *player(Types::ObjectID playerId);
 };